merge into backup main

+++ uef_diem_danh/Controllers/StudentController.cs
@@ -33,6 +33,17 @@ namespace uef_diem_danh.Controllers
             return View(students);
         }
 
+
+        [Authorize(Roles = "Admin")]
+        [HttpGet("tai-len-anh-hoc-vien")]
+        // Later refactor to get necessary fields
+        public IActionResult GetUploadStudentAvatarPage()
+        {
+
+            return View("~/Views/Student/UploadStudentAvatar.cshtml");
+        }
+
+
         [Route("api/lay-danh-sach-hoc-vien-theo-lop/{study_class_id}")]
         [HttpGet]
         public async Task<IActionResult> GetStudentsByStudyClass(int study_class_id)
@@ -62,12 +73,28 @@ namespace uef_diem_danh.Controllers
             }
         }
 
-        // Later refactor to get necessary fields
         [Route("api/lay-chi-tiet-hoc-vien/{student_id}")]
         [HttpGet]
         public async Task<IActionResult> GetDetailForUpdate(int student_id)
         {
-            HocVien studyClass = await context.HocViens.FindAsync(student_id);
+            //HocVien studyClass = await context.HocViens.FindAsync(student_id);
+
+            StudentDetailResponse studyClass = await context.HocViens
+                .Select(hv => new StudentDetailResponse
+                {
+                    MaHocVien = hv.MaHocVien,
+                    TenHinhAnh = hv.HinhAnh.Name,
+                    Ho = hv.Ho,
+                    Ten = hv.Ten,
+                    Email = hv.Email,
+                    DiaChi = hv.DiaChi,
+                    DonVi = hv.DonVi,
+                    MaBarCode = hv.MaBarCode,
+                    SoDienThoai = hv.SoDienThoai,
+                    NgaySinh = hv.NgaySinh
+                })
+                .FirstOrDefaultAsync(hv => hv.MaHocVien == student_id);
+
 
             return Ok(studyClass);
         }
@@ -182,6 +209,7 @@ namespace uef_diem_danh.Controllers
             try
             {
                 HocVien student = await context.HocViens
+                    .Include(hv => hv.HinhAnh)
                     .FirstOrDefaultAsync(lh => lh.MaHocVien == request.StudentId);
 
 
@@ -196,12 +224,15 @@ namespace uef_diem_danh.Controllers
 
                     // Find existing student avatar
                     string uploadFilePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "student_pictures");
-                    string existedStudentAvatarPath = Path.Combine(uploadFilePath, student.HinhAnh.Name);
 
-                    if (System.IO.File.Exists(existedStudentAvatarPath))
+                    if (context.HocViens.Any(hv => hv.HinhAnh.Name.Contains($"hv_{student.SoDienThoai}")))
                     {
-                        // Delete existed student avatar
-                        System.IO.File.Delete(existedStudentAvatarPath);
+                        string existedStudentAvatarPath = Path.Combine(uploadFilePath, student.HinhAnh.Name);
+                        if (System.IO.File.Exists(existedStudentAvatarPath))
+                        {
+                            // Delete existed student avatar
+                            System.IO.File.Delete(existedStudentAvatarPath);
+                        }
                     }
 
 
@@ -262,17 +293,20 @@ namespace uef_diem_danh.Controllers
             try
             {
                 HocVien student = await context.HocViens
+                    .Include(hv => hv.HinhAnh)
                     .FirstOrDefaultAsync(lh => lh.MaHocVien == request.MaHocVien);
 
 
                 // Find existing student avatar
                 string uploadFilePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "student_pictures");
-                string existedStudentAvatarPath = Path.Combine(uploadFilePath, student.HinhAnh.Name);
-
-                if (System.IO.File.Exists(existedStudentAvatarPath))
+                if (context.HocViens.Any(hv => hv.HinhAnh.Name.Contains($"hv_{student.SoDienThoai}")))
                 {
-                    // Delete existed student avatar
-                    System.IO.File.Delete(existedStudentAvatarPath);
+                    string existedStudentAvatarPath = Path.Combine(uploadFilePath, student.HinhAnh.Name);
+                    if (System.IO.File.Exists(existedStudentAvatarPath))
+                    {
+                        // Delete existed student avatar
+                        System.IO.File.Delete(existedStudentAvatarPath);
+                    }
                 }
 
 
@@ -309,8 +343,8 @@ namespace uef_diem_danh.Controllers
                 using var driver = new ChromeDriver(options);
 
 
-                //driver.Navigate().GoToUrl("http://127.0.0.1:5500/html/page/barcode-card-single.html");
                 driver.Navigate().GoToUrl($"https://laitsolution.id.vn/in-mot-the-hoc-vien/{student_id}");
+                //driver.Navigate().GoToUrl($"https://localhost:5046/in-mot-the-hoc-vien/{student_id}");
 
 
                 var printOptions = new PrintOptions
@@ -363,8 +397,9 @@ namespace uef_diem_danh.Controllers
                 options.AddArgument("--no-sandbox");
 
                 using var driver = new ChromeDriver(options);
-                //driver.Navigate().GoToUrl("http://127.0.0.1:5500/html/page/barcode-card-multiple.html");
+
                 driver.Navigate().GoToUrl($"https://laitsolution.id.vn/in-danh-sach-the-hoc-vien/{study_class_id}");
+                //driver.Navigate().GoToUrl($"https://localhost:5046/in-danh-sach-the-hoc-vien/{study_class_id}");
 
 
                 var printOptions = new PrintOptions
